fix: widen volume to BIGINT for A-share volumes + correct CSI 300 benchmark
- price_bars.volume: INTEGER → BIGINT (A-shares trade billions/day,
  overflowing 32-bit int); migration runs idempotently on startup
- LLM default benchmark for .SHH/.SHZ: 000300.SHH → 510300.SHH
  (CSI 300 ETF, actually fetchable from Alpha Vantage)

+++ apps/api/app/models/price_bar.py
@@ -1,6 +1,6 @@
 from datetime import date, datetime
 
-from sqlalchemy import Date, DateTime, Float, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, Date, DateTime, Float, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.session import Base
@@ -18,7 +18,7 @@ class PriceBar(Base):
     low: Mapped[float] = mapped_column(Float)
     close: Mapped[float] = mapped_column(Float)
     adjusted_close: Mapped[float] = mapped_column(Float)
-    volume: Mapped[int] = mapped_column(Integer)
+    volume: Mapped[int] = mapped_column(BigInteger)
     dividend_amount: Mapped[float] = mapped_column(Float, default=0.0)
     split_coefficient: Mapped[float] = mapped_column(Float, default=1.0)
     source: Mapped[str] = mapped_column(String(32), default="alpha_vantage")
